requirements.txt
extra_streamlit_components==0.1.60
mysql_connector_repackaged==0.3.1
numpy==1.26.2
pandas==2.1.4
PyJWT==2.8.0
PyJWT==2.8.0
pymongo==4.6.1
python_bcrypt==0.3.2
pytz==2023.3.post1
Requests==2.31.0
streamlit==1.28.2
streamlit_agraph==0.0.45
